Move all trip components inside page
This is in an effort to make everything seemless with adding Suspense.
